requirements.txt
tensorflow==2.19.0
keras==3.10.0
opencv-python
numpy
Pillow